app/requirements.txt
yt-dlp
openai-whisper
pydub
ffmpeg-python
ffprobe-python
whisper